% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolutionfun.R
\name{convolutionfun}
\alias{convolutionfun}
\title{Convolution function}
\usage{
convolutionfun(t, t0, f, g)
}
\arguments{
\item{t}{a vector of time points}

\item{t0}{initial time}

\item{f}{a function to convolve as f(t-u)}

\item{g}{a function to convolve}
}
\value{
a convolution function between f and g
}
\description{
Creates the convolution function between two functions f and g.
Instead of returning a numerical value, this function returns a function.
}
\examples{
tms<-seq(0,10, by=0.1)
cosconvfun<-convolutionfun(f=cos, g=cos) # convolve cosine function with itself
solfun<-function(t){ (t*cos(t) + sin(t))/2} # analytical solution (Braun 1993, Dif Eq and App, Springer, p. 254)
x1<-sapply(X=tms, FUN=cosconvfun)
x2<-sapply(X=tms, FUN=solfun)

plot(tms, x1, type="l")
lines(tms, x2, col=2)
}
