% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmanuelModel.R
\name{EmanuelModel}
\alias{EmanuelModel}
\title{Terrestrial carbon model of Emanuel}
\usage{
EmanuelModel()
}
\value{
a model of class lam
}
\description{
This function returns the model of Emanuel as an object of class lam.
It is mostly a constructor of the model to be used for further analysis.
}
\examples{
EmanuelModel()
}
