% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibriumOutflux.R
\name{equilibriumOutflux}
\alias{equilibriumOutflux}
\title{Equilibrium output flux for a linear autonomous model}
\usage{
equilibriumOutflux(model)
}
\arguments{
\item{model}{an object of class lam, a linear autonomous model}
}
\value{
a vector with the output fluxes for all compartments
}
\description{
Equilibrium output flux for a linear autonomous model
}
