% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Impulse response function}
\usage{
irf(model)
}
\arguments{
\item{model}{A model of class lam}
}
\value{
A function that takes time as main argument
}
\description{
Builds an impulse response function from a model object of class lam
}
\examples{
EmanuelIRF<-irf(model=EmanuelModel())
tms<-seq(0,100)
em<-sapply(tms, EmanuelIRF)
plot(tms, em, type="l")

}
