% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeB.R
\name{makeB}
\alias{makeB}
\title{Compartmental matrix from a set of prior parameters of Porce model}
\usage{
makeB(pars)
}
\arguments{
\item{pars}{a numeric vector of 16 parameter values}
}
\value{
A compartmental matrix of dimension 7
}
\description{
This function builds a compartmental matrix of seven pools using a set of
16 parameter values. It is mostly used to build a matrix from the parameter
values stored in the modpars dataset.
}
\examples{
makeB(pars=modpars[1,])
}
