% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modpars.R
\docType{data}
\name{modpars}
\alias{modpars}
\title{Model parameters for Porce}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 16 columns.
}
\usage{
data(modpars)
}
\description{
A dataset of model parameters that can be used as prior information for the Porce model.
The dataset corresponds to a seven pool model developed for the Porce region of
Colombia.
}
\references{
Sierra et al. (2021). Journal of Ecology 109(8): 2845–2855.
}
\keyword{datasets}
