% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks}
\alias{stocks}
\title{Time-dependent stocks for a linear autonomous model}
\usage{
stocks(t, t0, x0, model)
}
\arguments{
\item{t}{a scalar value of the time to obtain the solution of the model}

\item{t0}{a scalar value of the initial time}

\item{x0}{a vector containing the size of the compartments at the initial time}

\item{model}{an object of class lam, a linear autonomous model}
}
\value{
a vector with the stocks at the desired time
}
\description{
Time-dependent stocks for a linear autonomous model
}
