% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TECOModel.R
\name{TECOModel}
\alias{TECOModel}
\title{The Terrestrial Ecosystem Carbon model of Weng and Luo}
\usage{
TECOModel()
}
\value{
a model of class lam
}
\description{
This function returns the TECO model of Weng and Luo as an object of class lam.
It is mostly a constructor of the model to be used for further analysis.
}
\examples{
TECOModel()
}
