% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeB.R
\name{makeB}
\alias{makeB}
\title{Compartmental matrix from a vector of named parameters}
\usage{
makeB(pars)
}
\arguments{
\item{pars}{a numeric vector of appropriately named parameter values}
}
\value{
A compartmental matrix
}
\description{
This function builds a compartmental matrix using a vector of
parameter values. The parameter must be named using a name of the form "ki"
for rates for compartment i, and transfer coefficients "alphaij" with j the
number of the donor compartment and i the number of the receiver compartment.
}
\examples{
names(modpars[1,])
makeB(pars=modpars[1,])
}
